EcommersWebsiteVR_7f3d9/tests/test_main.py
import unittest
from unittest.mock import patch, MagicMock
from flask import Flask, render_template
from flask.testing import FlaskClient

# Mock implementation of Flask and its dependencies
class MockFlask:
    def __init__(self, name):
        self.name = name
        self.routes = {}

    def route(self, route, methods=None):
        def decorator(f):
            self.routes[route] = f
            return f
        return decorator

    def run(self, debug=False):
        pass

class MockRender:
    @staticmethod
    def render_template(template_name):
        return f"Rendered {template_name}"

# Implementation code
app = MockFlask(__name__)

@app.route('/')
def index():
    return MockRender.render_template('index.html')

@app.route('/toggle_navbar', methods=['POST'])
def toggle_navbar():
    return '', 204

if __name__ == '__main__':
    app.run(debug=True)

# Test suite
class TestFlaskApp(unittest.TestCase):
    def setUp(self):
        self.app = app
        self.client = FlaskClient(self.app)

    def test_index_route(self):
        response = self.app.routes['/']()
        self.assertEqual(response, "Rendered index.html")

    def test_toggle_navbar_route(self):
        response = self.app.routes['/toggle_navbar']()
        self.assertEqual(response, ('', 204))

    @patch('flask.render_template')
    def test_index_route_with_mock(self, mock_render_template):
        mock_render_template.return_value = "Mocked index.html"
        response = self.app.routes['/']()
        self.assertEqual(response, "Mocked index.html")
        mock_render_template.assert_called_once_with('index.html')

class TestClientSideJavaScript(unittest.TestCase):
    def setUp(self):
        self.js_code = """
        document.addEventListener('DOMContentLoaded', function() {
            const bar = document.getElementById('bar');
            const close = document.getElementById('close');
            const nav = document.getElementById('navbar');

            if (bar) {
                bar.addEventListener('click', function() {
                    nav.classList.add('active');
                });
            }

            if (close) {
                close.addEventListener('click', function() {
                    nav.classList.remove('active');
                });
            }
        });
        """

    def test_js_code_structure(self):
        self.assertIn('document.addEventListener', self.js_code)
        self.assertIn("getElementById('bar')", self.js_code)
        self.assertIn("getElementById('close')", self.js_code)
        self.assertIn("getElementById('navbar')", self.js_code)
        self.assertIn("classList.add('active')", self.js_code)
        self.assertIn("classList.remove('active')", self.js_code)

if __name__ == '__main__':
    unittest.main()